+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -13,8 +13,10 @@ import importlib  # allow forward references in type hints
 import json
 import logging
 import os.path
+import re
 import sys
 import threading
+from collections import defaultdict
 from dataclasses import dataclass
 from functools import reduce, wraps
 from ipaddress import IPv4Address
@@ -32,7 +34,7 @@ from ska_control_model import (
     SimulationMode,
     TestMode,
 )
-from ska_low_mccs_common import HealthRecorder, MccsBaseDevice
+from ska_low_mccs_common import MccsBaseDevice
 from ska_tango_base.base import CommandTracker
 from ska_tango_base.commands import (
     DeviceInitCommand,
@@ -69,11 +71,31 @@ from .firmware_threshold_interface import (
 )
 from .tile_component_manager import TileComponentManager
 from .tile_health_model import TileHealthModel
+from .tile_health_recorder import TileHealthRecorder
 from .tpm_status import TpmStatus
 
 __all__ = ["MccsTile", "main"]
 
 DevVarLongStringArrayType = tuple[list[ResultCode], list[str]]
+
+
+def is_v1(version_str: str) -> bool:
+    """
+    Return True if version is v1.x.x?, False if v2.x.x?.
+
+    :param version_str: version string to check.
+    :return: True if v1.x.x? (or empty string), False if v2.x.x?.
+    :raises ValueError: if version_str is not valid.
+    """
+    # Allow empty string (treat as v2 by default)
+    if version_str == "":
+        return False
+
+    match = re.fullmatch(r"v([12])\.\d+\.\d+[a-z]?", version_str)
+    if not match:
+        raise ValueError(f"Invalid version format: '{version_str}'")
+
+    return match.group(1) == "1"
 
 
 def engineering_mode_required(func: Callable) -> Callable:
@@ -221,7 +243,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         self._info: dict[str, Any] = {}
         self.component_manager: TileComponentManager
         self._stopping: bool
-        self._health_recorder: HealthRecorder | None
+        self._health_recorder: TileHealthRecorder | None
         self._health_report = ""
         self.hw_firmware_thresholds: FirmwareThresholds
         self.db_firmware_thresholds: FirmwareThresholds
@@ -232,6 +254,18 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         self.component_manager_fault: bool | None = None
         self.power_state: PowerState | None = None
         self.status_information: dict[str, str] = {}
+
+        self._intermediate_healths: dict[str, HealthState] = {
+            "temperatures": HealthState.UNKNOWN,
+            "currents": HealthState.UNKNOWN,
+            "dsp": HealthState.UNKNOWN,
+            "voltages": HealthState.UNKNOWN,
+            "io": HealthState.UNKNOWN,
+            "timing": HealthState.UNKNOWN,
+            "adcs": HealthState.UNKNOWN,
+            "alarms": HealthState.UNKNOWN,
+        }
+        self._intermediate_attrs: dict[str, list[str]]
 
     def delete_device(self: MccsTile) -> None:
         """
@@ -778,6 +812,12 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                 "FPGA1",
             ],
         }
+        self._intermediate_attrs = defaultdict(list)
+        for attr, path in self.attribute_monitoring_point_map.items():
+            if not path:
+                continue
+            self._intermediate_attrs[path[0]].append(attr.lower())
+
         super().init_device()
 
         self.db_firmware_thresholds = FirmwareThresholds()
@@ -848,6 +888,21 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                 self.push_change_event("healthState", health)
                 self.push_archive_event("healthState", health)
 
+    def _intermediate_health_changed(
+        self: MccsTile,
+        group: str,
+        health: HealthState,
+    ) -> None:
+        if self._stopping:
+            return
+
+        if self.UseAttributesForHealth:
+            if self._intermediate_healths[group] != health:
+                self.logger.info(
+                    "Intermediate Health changed ==> " f"{group=} {health=}"
+                )
+                self._intermediate_healths[group] = health
+
     def _attr_conf_changed(self: MccsTile, attribute_name: str) -> None:
         """
         Handle change in attribute configuration.
@@ -886,12 +941,24 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
                 "coreCommunicationStatus",
             }
 
-            self._health_recorder = HealthRecorder(
+            if is_v1(self.HardwareVersion):
+                # Older tiles do not have ADC temperature sensors.
+                for idx in range(16):
+                    healthful_attrs.discard(f"temperatureADC{idx}")
+
+            if self.BiosVersion == "0.5.0":
+                # This monitoring point was introduced in 0.6.0.
+                # If we are bios 0.5.0 we ignore it.
+                healthful_attrs.discard("timing_pll_40g_count")
+
+            self._health_recorder = TileHealthRecorder(
                 self.get_name(),
                 logger=self.logger,
                 attributes=list(healthful_attrs),
                 health_callback=self._health_changed_new,
                 attr_conf_callback=self._attr_conf_changed,
+                intermediate_health_callback=self._intermediate_health_changed,
+                health_groups=self._intermediate_attrs,
             )
         else:
             self._health_model = TileHealthModel(
@@ -4367,12 +4434,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: temperature Health State of the device
-
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["temperatures"]
         return self._health_model.intermediate_healths["temperatures"][0]
 
     @attribute(dtype=HealthState)
@@ -4385,12 +4449,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: voltage Health State of the device
-
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["voltages"]
         return self._health_model.intermediate_healths["voltages"][0]
 
     @attribute(dtype=HealthState)
@@ -4403,11 +4464,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: current Health State of the device
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["currents"]
         return self._health_model.intermediate_healths["currents"][0]
 
     @attribute(dtype=HealthState)
@@ -4420,11 +4479,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: alarm Health State of the device
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["alarms"]
         return self._health_model.intermediate_healths["alarms"][0]
 
     @attribute(dtype=HealthState)
@@ -4437,11 +4494,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: ADC Health State of the device
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["adcs"]
         return self._health_model.intermediate_healths["adcs"][0]
 
     @attribute(dtype=HealthState)
@@ -4454,11 +4509,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: timing Health State of the device
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["timing"]
         return self._health_model.intermediate_healths["timing"][0]
 
     @attribute(dtype=HealthState)
@@ -4471,11 +4524,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: io Health State of the device
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["io"]
         return self._health_model.intermediate_healths["io"][0]
 
     @attribute(dtype=HealthState)
@@ -4488,11 +4539,9 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         the overall healthState of the tile.
 
         :return: dsp Health State of the device
-        :raises NotImplementedError: If UseAttributesForHealth
-            if True
         """
         if self.UseAttributesForHealth:
-            raise NotImplementedError("")
+            return self._intermediate_healths["dsp"]
         return self._health_model.intermediate_healths["dsp"][0]
 
     @attribute(dtype="DevString")